requirements.txt
streamlit==1.47.1
openai>=1.98.0
python-dotenv==1.1.1